+++ src/connected/StatisticsPanel.tsx
@@ -13,6 +13,7 @@ import {
   selectedDatasetSelector,
   selectedDatasetTimeLabelSelector,
   selectedPlaceInfoSelector,
+  selectedVariableDimensionValuesSelector,
   selectedVariableSelector,
 } from "@/selectors/controlSelectors";
 import _StatisticsPanel from "@/components/StatisticsPanel";
@@ -25,6 +26,7 @@ const mapStateToProps = (state: AppState) => {
     selectedDataset: selectedDatasetSelector(state),
     selectedVariable: selectedVariableSelector(state),
     selectedTime: selectedDatasetTimeLabelSelector(state),
+    selectedDimensionValues: selectedVariableDimensionValuesSelector(state),
     selectedPlaceInfo: selectedPlaceInfoSelector(state),
     statisticsLoading: statisticsLoadingSelector(state),
     statisticsRecords: resolvedStatisticsRecordsSelector(state),