+++ src/states/controlState.ts
@@ -23,8 +23,8 @@ import { loadUserSettings } from "./userSettings";
 import { PaletteMode } from "@mui/material";
 import { LayerState } from "@/model/layerState";
 
-export type TimeAnimationInterval = 250 | 500 | 1000 | 2500;
-export const TIME_ANIMATION_INTERVALS: TimeAnimationInterval[] = [
+export type DimensionAnimationInterval = 250 | 500 | 1000 | 2500;
+export const DIMENSION_ANIMATION_INTERVALS: DimensionAnimationInterval[] = [
   250, 500, 1000, 2500,
 ];
 
@@ -66,6 +66,8 @@ export type LayerGroupStates = {
   baseMaps?: boolean;
 };
 
+export type DimensionValues = Record<string, string | number | null>;
+
 export type ThemeMode = PaletteMode | "system";
 export const THEME_NAMES: ThemeMode[] = ["light", "dark", "system"];
 export const THEME_LABELS: [ThemeMode, string][] = [
@@ -93,7 +95,6 @@ export interface ControlState {
   selectedTimeRange: TimeRange | null;
   timeSeriesUpdateMode: "add" | "replace";
   timeAnimationActive: boolean;
-  timeAnimationInterval: TimeAnimationInterval;
   timeChunkSize: number;
   autoShowTimeSeries: boolean;
   timeSeriesChartTypeDefault: TimeSeriesChartType;
@@ -140,6 +141,11 @@ export interface ControlState {
   showZoomInfoBox: boolean;
   zoomLevel: number | undefined;
   datasetZLevel: number | undefined;
+  dimensionAnimationInterval: DimensionAnimationInterval;
+  dimensionAnimationActive: boolean;
+  selectedDimensionLabel: string | null;
+  selectedDimensionValues: DimensionValues;
+  showAllDimensions: boolean;
 }
 
 export function newControlState(): ControlState {
@@ -157,7 +163,6 @@ export function newControlState(): ControlState {
     selectedTimeRange: null,
     timeSeriesUpdateMode: "add",
     timeAnimationActive: false,
-    timeAnimationInterval: 1000,
     timeChunkSize: 20,
     autoShowTimeSeries: true,
     timeSeriesChartTypeDefault: "line",
@@ -225,6 +230,11 @@ export function newControlState(): ControlState {
     showZoomInfoBox: branding.showZoomInfoBox || false,
     zoomLevel: undefined,
     datasetZLevel: undefined,
+    dimensionAnimationInterval: 1000,
+    dimensionAnimationActive: false,
+    selectedDimensionLabel: null,
+    selectedDimensionValues: {},
+    showAllDimensions: false,
   };
   return loadUserSettings(state);
 }